refactored tripleextractor
