+++ rdf/extractor/TripleExtractor.py
@@ -1,5 +1,6 @@
 from __future__ import annotations
 import datetime
+from abc import abstractmethod
 from typing import List, OrderedDict, Any, Tuple, NamedTuple
 
 import spacy
@@ -34,15 +35,9 @@ class TripleExtractor:
         """
         # Extract publication info and adds it to the RDF triples.
         self.extract_publication(document)
-
-        for article in document.articles:
-            # For each article, process the text and extract non-textual data in it.
-            self.__process_article(article)
-            self.__extract_article(article, document)
-
+        self.extract_content(document)
         # Adds named individuals to the triples list.
         self.__append_named_individual()
-
         # Function from rdf.RdfCreator, writes triples to file
         store_rdf_triples(self.triples)
 
@@ -140,6 +135,10 @@ class TripleExtractor:
                 generate_relation(RelationTypeConstants.RDF_TYPE),
                 generate_uri_reference(self.namespace, ref=prop2)
             ))
+
+    @abstractmethod
+    def extract_content(self, document: Document):
+        pass
 
 
 class Triple(NamedTuple):
