+++ rdf/extractor/__init__.py
@@ -1,4 +1,4 @@
 from .TripleExtractorEnum import TripleExtractorEnum
 from .NJTripleExtractor import NJTripleExtractor
 from .GFTripleExtractor import GFTripleExtractor
-from .TripleExtractor import TripleExtractor
+from .TripleExtractor import TripleExtractor, Triple
